Crate pallet_evm_precompile_randomness

Precompile to interact with randomness through an evm precompile.

Structs

RandomnessPrecompile

A precompile to wrap the functionality from pallet-randomness

Enums

RandomnessPrecompileCall

Constants

EXECUTE_EXPIRATION_ESTIMATED_COST

INCREASE_REQUEST_FEE_ESTIMATED_COST

LOG_FULFILLMENT_FAILED

LOG_FULFILLMENT_SUCCEEDED

REQUEST_RANDOMNESS_ESTIMATED_COST

Functions

log_fulfillment_failed

log_fulfillment_succeeded

prepare_and_finish_fulfillment_gas_cost

Fulfillment overhead cost, which takes input weight hint -> weight -> return gas

subcall_overhead_gas_costs

transaction_gas_refund